Cleanup: extract components into different files
