+++ lib/vyasa_web/live/display_manager/components/comment_binding.ex
@@ -1,0 +1,19 @@
+defmodule VyasaWeb.DisplayManager.Components.CommentBinding do
+  use VyasaWeb, :html
+
+  def render(assigns) do
+    assigns = assigns |> assign(:elem_id, "comment-modal-#{Ecto.UUID.generate()}")
+
+    ~H"""
+    <span
+      :for={comment <- @comments}
+      class="block
+                 before:content-['╰'] before:mr-1 before:text-gray-500
+                 lg:before:content-none
+                 lg:border-l-0 lg:pl-2"
+    >
+      <%= comment.body %> - <b><%= comment.signature %></b>
+    </span>
+    """
+  end
+end
